+++ src/database/migrations/0027-alter-organizations-rename-ein.sql
@@ -1,0 +1,2 @@
+ALTER TABLE organizations
+  RENAME COLUMN employer_identification_number TO tax_id;