+++ src/database/queries/organizations/selectByTaxId.sql
@@ -1,3 +1,3 @@
 SELECT organization_to_json(organizations) AS "object"
 FROM organizations
-WHERE employer_identification_number = :employerIdentificationNumber
+WHERE tax_id = :taxId