+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - The GET `/opportunities` endpoint now returns an `OpportunityBundle` and accepts pagination parameters.
+- Added a new `Organization` attribute `taxId` which contains the same value as `employerIdentificationNumber`.
+- Deprecated the `Organizations` attribute `employerIdentificationNumber`. Please update your clients to use `taxId` instead.
 
 ## 0.8.0 2024-04-23